requirements.txt
# PyTorch (Meta / Linux Foundation) - install ROCm build separately:
#   pip install torch torchvision --index-url https://download.pytorch.org/whl/rocm7.2
# Uncomment below for CUDA or CPU-only builds:
# torch>=2.4.0
# torchvision>=0.19.0

# Hugging Face ecosystem
accelerate>=0.30.0
transformers>=4.40.0
safetensors>=0.4.0
datasets>=2.14.0
kagglehub>=0.3.0

# Core scientific computing (NumPy, SciPy, scikit-learn)
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0

# Image processing
Pillow>=10.0.0
opencv-python-headless>=4.8.0

# Networking
websockets>=14.0
requests>=2.31.0

# Utilities
einops>=0.7.0
tqdm>=4.66.0
pyyaml>=6.0

# Perceptual loss
lpips>=0.1.4

# Evaluation
pytorch-fid>=0.3.0
